Git: remove & ignore generated files
These are generated by the CPSA build and other processes:

tst/*.{txt,xhtml)
doc/examples/*.xhtml
dist-newstyle/

It's confusing to have Git notice these when they're created, or to
have them actually recorded in the repository, because it makes it
look to the user that they've made changes to the source which ought
to be committed. (Indeed, that's often how files like these get
mistakenly added to begin with, even if it was never intended.)

This commit removes any of these files which are in
the repository now, and adds configuration so that Git will ignore all
of them.

If you want to include the sample output from the tutorial examples
(doc/examples/*.xhtml), I suggest putting these in a separate
directory, so they aren't confused with and modified by output
produced merely by following the tutorial, or using "make clean"
there.